Implemented Wall Removal, Rotating Wall and New Scoring System
Implemented the rotating wall, so after 5 throws and up til the 11th throw, pieces of the wall are removed. Then after the 10th throw and on, the wall starts rotating.
Implemented new scoring system so that the farther the player gets in the round, the more exponential the score multiplier becomes. Also implemented a "real-time" score multiplier update to the UI in game.