Fixed the weird ball bounces when hit off the side wall
Fixed issue where the ball would bounce back towards the side wall when the ball hit the side wall first.
Still need to make the apply to the rest of the right wall, and to the entire left wall.
Also need to make the force apporpriate for each level of speed.