Tweaked Cube movement settings
Changed the way the cubes move so that they don't go too slow or too fast, while also attempting to minimize how often cubes stop after hitting each other.

Applied textures to wall, but need updating.
Changed the side walls to be 90 degrees
Updated the scoring system so that each wall hit adds .25 to the multiplier as well as when the player catches the ball off the back wall